feat(one-adk)!: delete the legacy Gemini relay and client-side lexical planner
Clean cutover: One's ADK agent tree is now the only decision-maker.

Backend:
- extract relay ticket auth into api/routes/one/relay_auth.py (same signed
  v1 HMAC one-time-nonce contract, minus persona-hint plumbing)
- delete api/routes/kai/agent_realtime_gemini.py (hand-rolled Vertex pump,
  ephemeral token endpoint, action-proposal extraction) + its tests and
  router registration; /api/kai/agent/realtime/gemini/* now 404s

Frontend:
- delete lib/voice/one-voice-live-action-bridge.ts (client-side lexical
  re-ranker lane that raced Gemini's proposals - the root cause of the
  'random commands' bug) + its tests and the intelligence eval built on it
- remove action_proposal from the transport event union, the proposal
  normalizer from gemini-live-client, fetchGeminiLiveToken, and the
  agent-bar planner wiring; transcripts now only mirror into the session

Verified: backend test-ci 272 passed, server route table has only
/api/one/adk/*; live smoke: new relay speaks, legacy endpoint 404; webapp
1693 tests passed, typecheck + eslint + design-system clean; surface map
regenerated with no diff.
